Refactored to fit the requirements
